refactor: move mock passports data to data module
