+++ src/data/index.ts
@@ -1,3 +1,4 @@
 export * from './mockDpps';
 export * from './simpleMockProducts';
 export * from './mockSuppliers';
+export * from './mockPublicPassports';
